requirements.txt
asgiref==3.6.0
asn1crypto==0.24.0
astroid==2.2.5
autopep8==1.4.4
bcrypt==3.1.7
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cssselect2==0.7.0
Django==2.2.17
django-appconf==1.0.3
django-bower==5.2.0
django-classy-tags==3.0.1
django-cms==3.9.0
django-compat==1.0.15
django-compressor==2.3
django-extensions==2.2.1
django-filer==2.1.2
django-formtools==2.1
django-js-asset==1.2.2
django-mptt==0.10.0
django-polymorphic==2.1.2
django-recaptcha==3.0.0
django-sekizai==3.0.1
Django-Select2==4.3.2
django-treebeard==4.3
djangocms-admin-style==1.4.0
djangocms-attributes-field==1.1.0
djangocms-file==2.3.0
djangocms-googlemap==2.0.0
djangocms-link==2.5.0
djangocms-picture==2.3.0
djangocms-snippet==3.1.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
djangorestframework==3.8.0
easy-thumbnails==2.8.1
Fabric==1.7.0
flup==1.0.3
gevent==21.1.2
greenlet==1.1.0
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
invoke==1.3.0
isort==4.3.21
lazy-object-proxy==1.4.2
lxml==4.9.2
mccabe==0.6.1
olefile==0.46
paramiko==2.6.0
Pillow==8.2.0
psycopg2==2.9.1
pycodestyle==2.5.0
pycparser==2.19
pygeocoder==1.2.5
pylint==2.3.1
PyNaCl==1.3.0
pytz==2019.2
rcssmin==1.0.6
reportlab==3.6.9
requests==2.22.0
rjsmin==1.1.0
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typed-ast==1.4.3
Unidecode==1.0.23
urllib3==1.25.3
webencodings==0.5.1
Werkzeug==0.15.6
wrapt==1.11.2
zope.event==4.5.0
zope.interface==5.4.0
